fix: core — 20 compilation errors across 5 root causes
P0-1: OvershootHelper.resetOvershoot() — moved from inside object to
  top-level extension (was inaccessible to LegitAimStrategy/SelfAdaptive).

P0-2: AimStrategy.State — made 'open class' so AdaptiveState can extend.
  AdaptiveState now overrides reset() to clear alignmentEma etc.

P1-1: ProbabilisticClickStrategy — added missing import ClickMode.

P1-2: LegitVelocityStrategy.processLegit — added state parameter
  (used state.delayQueue/state.tickCounter but never received state).

P2: GaussianNoise.ln(s) → ln(s.toDouble()), RotationCalculator
  bestAngularDist = angularDist.toDouble() (Float→Double mismatch).

+++ SwitchLite/core/src/main/kotlin/io/switchlite/core/algorithm/RotationCalculator.kt
@@ -148,7 +148,7 @@ object RotationCalculator {
             val pitchDiff = kotlin.math.abs(rot.pitch - currentAim.pitch)
             val angularDist = yawDiff + pitchDiff
             if (angularDist < bestAngularDist) {
-                bestAngularDist = angularDist
+                bestAngularDist = angularDist.toDouble()
                 bestPoint = closest
             }
         }
